changed modules directory

+++ Makefile
@@ -5,17 +5,16 @@ all: install
 install:
 	mkdir -p ~/.config/bdwmb
 	mkdir -p ~/.config/bdwmb/modules
-	cp config.sh ~/.config/bdwmb
+	if [ ! -f ~/.config/bdwmb/config.sh ]; then cp config.sh ~/.config/bdwmb; fi
 
 uninstall:
-	rm -rf ~/.config/bdwmb
-	${RUNAS} rm -rf /usr/local/bin/bdwmb /usr/local/bin/modules
+	${RUNAS} rm -rf /usr/local/bin/bdwmb /usr/local/bin/bdwmb_modules
 
 bin:
 	mkdir -p /usr/local/bin
 	cp bdwmb.sh /usr/local/bin/bdwmb
 	chmod a+rx /usr/local/bin/bdwmb
-	cp -r modules /usr/local/bin
+	cp -r bdwmb_modules /usr/local/bin
 
 full:
 	make uninstall
